Move upload logic to separate object
